move file operations from b3sum to blake3
