+++ Cargo.toml
@@ -79,6 +79,8 @@ no_neon = []
 
 zeroize = ["zeroize_crate", "arrayvec/zeroize"]
 
+file = ["memmap2", "rayon", "std"]
+
 [package.metadata.docs.rs]
 # Document Hasher::update_rayon on docs.rs.
 features = ["rayon"]
@@ -91,6 +93,7 @@ rayon = { version = "1.2.1", optional = true }
 cfg-if = "1.0.0"
 digest = { version = "0.10.1", features = [ "mac" ], optional = true }
 zeroize_crate = { package = "zeroize", version = "1", default-features = false, features = ["zeroize_derive"], optional = true }
+memmap2 = { version = "0.7.1", optional = true }
 
 [dev-dependencies]
 hex = "0.4.2"
